Develop

+++ app/Http/Controllers/Admin/Product/DestroyController.php
@@ -10,7 +10,10 @@ class DestroyController extends Controller
 {
     public function __invoke(Product $product)
     {
+        $this->authorize('delete', $product);
+
         $product->delete();
-        return response()->json('Product removed', 200);
+
+        return response()->json('Продукт успешно удален', 200);
     }
 }
